Fixed conflict between field and class property.
The django template `{% fieldset.XYZ %}` tries to be intelligent and
checks whether `fieldset['XYZ']` exists before checking `fieldset.XYZ`.
This can cause issues when trying to fetch a class variable with a
conflicting field name.

In particular, if a field is called `name`, then `{% fieldset.name %}` will
fetch the field called `name` as opposed to the fieldset's name.

Signed-off-by: JP-Ellis <coujellis@gmail.com>